backend/requirements.txt
asgiref==3.11.0
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssbeautifier==1.15.4
curl_cffi==0.14.0
dj-database-url==2.3.0
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djlint==1.36.4
EditorConfig==0.17.1
elementpath==5.1.1
enum-tools==0.12.0
fints==5.0.0
ibflex==0.15
idna==3.11
jsbeautifier==1.15.4
json5==0.13.0
lxml==6.0.2
mt-940==4.30.0
pathspec==1.0.4
psycopg2-binary==2.9.10
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
sepaxml==2.7.0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.6.3
webencodings==0.5.1
xmlschema==4.3.1
